Remove warnings, errors during host tests in CI
Debug strings often included format parameters which did not exactly match
the passed in format parameters, resulting in warnings in the host test build
process like
````
/home/runner/work/Arduino/Arduino/libraries/ESP8266WiFi/src/ESP8266WiFiAP.cpp:107:20: warning: format ‘%u’ expects argument of type ‘unsigned int’, but argument 3 has type ‘size_t’ {aka ‘long unsigned int’} [-Wformat=]
  107 |         DEBUG_WIFI("[AP] SSID length %u, too long or missing!\n", ssid_len);
      |                                                                   ~~~~~~~~
      |                                                                   |
      |                                                                   size_t {aka long unsigned int}
````

Fix by applying casting or PRxxx macros as appropriate.

Also, fix one debug message which was trying to use a `String` as a `char *`:
````
/home/runner/work/Arduino/Arduino/libraries/ESP8266WiFi/src/ESP8266WiFiMulti.cpp: In member function ‘wl_status_t ESP8266WiFiMulti::connectWiFiMulti(uint32_t)’:
/home/runner/work/Arduino/Arduino/libraries/ESP8266WiFi/src/ESP8266WiFiMulti.cpp:331:34: warning: format ‘%s’ expects argument of type ‘char*’, but argument 3 has type ‘String’ [-Wformat=]
  331 |                 DEBUG_WIFI_MULTI("[WIFIM] Connecting %s\n", ssid);
````

+++ libraries/ESP8266WiFi/src/ESP8266WiFi.cpp
@@ -75,12 +75,12 @@ void ESP8266WiFiClass::printDiag(Print& p) {
     char ssid[33]; //ssid can be up to 32chars, => plus null term
     memcpy(ssid, conf.ssid, sizeof(conf.ssid));
     ssid[32] = 0; //nullterm in case of 32 char ssid
-    p.printf_P(PSTR("SSID (%d): %s\n"), strlen(ssid), ssid);
+    p.printf_P(PSTR("SSID (%u): %s\n"), (unsigned int)strlen(ssid), ssid);
 
     char passphrase[65];
     memcpy(passphrase, conf.password, sizeof(conf.password));
     passphrase[64] = 0;
-    p.printf_P(PSTR("Passphrase (%d): %s\n"), strlen(passphrase), passphrase);
+    p.printf_P(PSTR("Passphrase (%u): %s\n"), (unsigned int)strlen(passphrase), passphrase);
 
     p.print(F("BSSID set: "));
     p.println(conf.bssid_set);
